backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.0.1

# Utilities
pydantic==2.5.0
python-jose[cryptography]==3.3.0

# Optional: For model serving and deployment
gunicorn==21.2.0

# Development dependencies (optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0
